src/com/leetcode/ValidParenthesis.java
package com.leetcode;

import java.util.Stack;

/**
 * Created by gpitois on 1/8/18.
 */
public class ValidParenthesis {
    public boolean isValid(String s) {
        Stack<Character> stack = new Stack<>();
        for (int i = 0; i < s.length(); i++) {
            switch (s.charAt(i)) {
                case '(':
                case '{':
                case '[':
                    stack.push(s.charAt(i));
                    break;
                case ')':
                    if (stack.empty() || stack.pop() != '(') return false;
                    break;
                case '}':
                    if (stack.empty() || stack.pop() != '{') return false;
                    break;
                case ']':
                    if (stack.empty() || stack.pop() != '[') return false;
                    break;
                default:
                    return false;
            }
        }
        return stack.empty();
    }
}
